\name{DataCpp-package}
\alias{DataCpp-package}
\alias{DataCpp}
\docType{package}
\title{
\packageTitle{DataCpp}
}
\description{
\packageDescription{DataCpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DataCpp}
\packageIndices{DataCpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DataCpp}

Maintainer: \packageMaintainer{DataCpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
